BitArray now supports 32bit as max elements.

+++ include/dm/datastructures/bitarray_inline_impl.h
@@ -3,58 +3,58 @@
  * License: http://www.opensource.org/licenses/BSD-2-Clause
  */
 
-void set(uint16_t _bit)
+void set(uint32_t _bit)
 {
     DM_CHECK(_bit < m_max, "bitArraySet | %d, %d", _bit, m_max);
 
-    const uint16_t bucket = _bit>>6;
+    const uint32_t bucket = _bit>>6;
     const uint64_t bit    = UINT64_C(1)<<(_bit&63);
     m_bits[bucket] |= bit;
 }
 
-void unset(uint16_t _bit)
+void unset(uint32_t _bit)
 {
     DM_CHECK(_bit < max(), "bitArrayUnset | %d, %d", _bit, max());
 
-    const uint16_t bucket = _bit>>6;
+    const uint32_t bucket = _bit>>6;
     const uint64_t bit    = UINT64_C(1)<<(_bit&63);
     m_bits[bucket] &= ~bit;
 }
 
-void toggle(uint16_t _bit)
+void toggle(uint32_t _bit)
 {
     DM_CHECK(_bit < max(), "bitArrayToggle | %d, %d", _bit, max());
 
-    const uint16_t bucket = _bit>>6;
+    const uint32_t bucket = _bit>>6;
     const uint64_t bit    = UINT64_C(1)<<(_bit&63);
     m_bits[bucket] ^= bit;
 }
 
-bool isSet(uint16_t _bit)
+bool isSet(uint32_t _bit)
 {
     DM_CHECK(_bit < max(), "bitArrayIsSet | %d, %d", _bit, max());
 
-    const uint16_t bucket = _bit>>6;
+    const uint32_t bucket = _bit>>6;
     const uint64_t bit    = UINT64_C(1)<<(_bit&63);
     return (0 != (m_bits[bucket] & bit));
 }
 
 private:
-inline uint16_t setRightmostBit(uint16_t _slot)
+inline uint32_t setRightmostBit(uint32_t _slot)
 {
     const uint64_t rightmost = m_bits[_slot]+1;
     m_bits[_slot] |= rightmost;
 
-    const uint16_t pos = uint16_t(bx::uint64_cnttz(rightmost));
-    const uint16_t idx = (_slot<<6)+pos;
-    const uint16_t max = this->max();
+    const uint32_t pos = uint32_t(bx::uint64_cnttz(rightmost));
+    const uint32_t idx = (_slot<<6)+pos;
+    const uint32_t max = this->max();
     return idx < max ? idx : max;
 }
 public:
 
-uint16_t setFirst()
+uint32_t setFirst()
 {
-    for (uint16_t slot = 0, end = numSlots(); slot < end; ++slot)
+    for (uint32_t slot = 0, end = numSlots(); slot < end; ++slot)
     {
         const bool hasUnsetBits = (m_bits[slot] != UINT64_MAX);
         if (hasUnsetBits)
@@ -66,12 +66,12 @@ uint16_t setFirst()
     return max();
 }
 
-uint16_t setAny()
+uint32_t setAny()
 {
-    const uint16_t begin = m_last;
-    const uint16_t count = numSlots();
+    const uint32_t begin = m_last;
+    const uint32_t count = numSlots();
 
-    for (uint16_t slot = begin, end = numSlots(); slot < end; ++slot)
+    for (uint32_t slot = begin, end = numSlots(); slot < end; ++slot)
     {
         const bool hasUnsetBits = (m_bits[slot] != UINT64_MAX);
         if (hasUnsetBits)
@@ -86,7 +86,7 @@ uint16_t setAny()
 
     m_last = (m_last >= count) ? 0 : m_last;
 
-    for (uint16_t slot = 0, end = begin; slot < end; ++slot)
+    for (uint32_t slot = 0, end = begin; slot < end; ++slot)
     {
         const bool hasUnsetBits = (m_bits[slot] != UINT64_MAX);
         if (hasUnsetBits)
@@ -103,14 +103,14 @@ uint16_t setAny()
 }
 
 /// Returns max() if none set.
-uint16_t getFirstSetBit()
+uint32_t getFirstSetBit()
 {
-    for (uint16_t ii = 0, end = numSlots(); ii < end; ++ii)
+    for (uint32_t ii = 0, end = numSlots(); ii < end; ++ii)
     {
         const bool hasSetBits = (0 != m_bits[ii]);
         if (hasSetBits)
         {
-            const uint16_t pos = uint16_t(bx::uint64_cnttz(m_bits[ii]));
+            const uint32_t pos = uint32_t(bx::uint64_cnttz(m_bits[ii]));
             return (ii<<6)+pos;
         }
     }
@@ -118,15 +118,15 @@ uint16_t getFirstSetBit()
     return max();
 }
 
-uint16_t getFirstUnsetBit()
+uint32_t getFirstUnsetBit()
 {
-    for (uint16_t ii = 0, end = numSlots(); ii < end; ++ii)
+    for (uint32_t ii = 0, end = numSlots(); ii < end; ++ii)
     {
         const bool hasUnsetBits = (m_bits[ii] != UINT64_MAX);
         if (hasUnsetBits)
         {
             const uint64_t sel = markFirstUnsetBit(m_bits[ii]);
-            const uint16_t pos = uint16_t(bx::uint64_cnttz(sel));
+            const uint32_t pos = uint32_t(bx::uint64_cnttz(sel));
             return (ii<<6)+pos;
         }
     }
@@ -135,9 +135,9 @@ uint16_t getFirstUnsetBit()
 }
 
 /// Returns max() if none set.
-uint16_t getLastSetBit()
+uint32_t getLastSetBit()
 {
-    for (uint16_t ii = numSlots(); ii--; )
+    for (uint32_t ii = numSlots(); ii--; )
     {
         const bool hasSetBits = (0 != m_bits[ii]);
         if (hasSetBits)
@@ -151,7 +151,7 @@ uint16_t getLastSetBit()
             {
                 const uint64_t sel = markFirstUnsetBit(m_bits[ii]);
                 const uint64_t leading = bx::uint64_cntlz(sel);
-                const uint16_t pos = 63-uint16_t(leading);
+                const uint32_t pos = 63-uint32_t(leading);
                 return ((ii)<<6)+pos;
             }
         }
@@ -160,15 +160,15 @@ uint16_t getLastSetBit()
     return 0;
 }
 
-uint16_t getLastUnsetBit()
+uint32_t getLastUnsetBit()
 {
-    for (uint16_t ii = numSlots(); ii--; )
+    for (uint32_t ii = numSlots(); ii--; )
     {
         const bool hasSetBits = (0 != m_bits[ii]);
         if (hasSetBits)
         {
             const uint64_t leading = bx::uint64_cntlz(m_bits[ii]);
-            const uint16_t pos = 63-uint16_t(leading);
+            const uint32_t pos = 63-uint32_t(leading);
             return (ii<<6)+pos;
         }
     }
@@ -176,16 +176,16 @@ uint16_t getLastUnsetBit()
     return max();
 }
 
-uint16_t count()
+uint32_t count()
 {
     uint64_t count = 0;
-    for (uint16_t ii = numSlots(); ii--; )
+    for (uint32_t ii = numSlots(); ii--; )
     {
         const uint64_t curr = m_bits[ii];
         count += bx::uint64_cntbits(curr);
     }
 
-    return uint16_t(count);
+    return uint32_t(count);
 }
 
 void reset()
